feat: EventType 추가
