+++ src/main/java/hibernate/event/delete/DeleteEventListener.java
@@ -1,4 +1,4 @@
-package hibernate.event.listener;
+package hibernate.event.delete;
 
 public interface DeleteEventListener {
 }
