+++ src/main/java/hibernate/event/load/LoadEventListener.java
@@ -1,4 +1,4 @@
-package hibernate.event.listener;
+package hibernate.event.load;
 
 public interface LoadEventListener {
 }
