+++ src/test/java/hibernate/event/delete/DeleteEventTest.java
@@ -1,4 +1,4 @@
-package hibernate.event.listener;
+package hibernate.event.delete;
 
 import hibernate.metamodel.BasicMetaModel;
 import hibernate.metamodel.MetaModel;
@@ -7,12 +7,10 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import org.junit.jupiter.api.Test;
 
-import java.util.Map;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-class MergeEventTest {
+class DeleteEventTest {
 
     private final MetaModel metamodel = MetaModelImpl.createPackageMetaModel(
             BasicMetaModel.createPackageMetaModel("hibernate.event.listener"),
@@ -20,12 +18,12 @@ class MergeEventTest {
     );
 
     @Test
-    void MergeEvent를_생성한다() {
-        MergeEvent actual = MergeEvent.createEvent(metamodel, TestEntity.class, 1L, Map.of());
+    void DeleteEvent를_생성한다() {
+        TestEntity givenEntity = new TestEntity();
+        DeleteEvent actual = DeleteEvent.createEvent(metamodel, givenEntity);
         assertAll(
-                () -> assertThat(actual.getEntityPersister()).isNotNull(),
-                () -> assertThat(actual.getEntityId()).isEqualTo(1L),
-                () -> assertThat(actual.getChangeColumns()).isNotNull()
+                () -> assertThat(actual.getEntity()).isEqualTo(givenEntity),
+                () -> assertThat(actual.getEntityPersister()).isNotNull()
         );
     }
 
@@ -33,7 +31,5 @@ class MergeEventTest {
     private static class TestEntity {
         @Id
         private Long id;
-
-        private String name;
     }
 }
